Split clock estimator source
