+++ CMakeLists.txt
@@ -31,6 +31,7 @@ add_executable(urg_stamped
   src/device_state_estimator.cpp
   src/scan_estimator_utm.cpp
   src/scan_estimator_ust.cpp
+  src/clock_estimator_uust1.cpp
   src/ros_logger.cpp
   src/scip2/logger.cpp
   src/urg_stamped.cpp
