+++ test/CMakeLists.txt
@@ -3,6 +3,7 @@ catkin_add_gtest(test_device_state_estimator
   ../src/device_state_estimator.cpp
   ../src/scan_estimator_utm.cpp
   ../src/scan_estimator_ust.cpp
+  ../src/clock_estimator_uust1.cpp
   ../src/scip2/logger.cpp
 )
 target_link_libraries(test_device_state_estimator ${catkin_LIBRARIES} ${Boost_LIBRARIES})
